refactor!: move config annotations to own package
This will break anything that references @ConfigFile.

This has been done to prepare for the new configuration features
